+++ examples/prime-sum-benchmark.cpp
@@ -4,8 +4,8 @@
 // see for licence details (MIT) at its official distribution:
 // https://github.com/FOLF-projects/folf-namespace
 // ===========================================================
-#include <folf-namespace/folf.hpp>
-#include <folf-namespace/folf-info.hpp>
+#include <folf/folf.hpp>
+#include <folf/folf-info.hpp>
 #include <iostream>
 #include <thread>
