+++ folf/algorithms.cpp
