+++ folf/consoleOperations.cpp
